centralize logic
